Fix compilation warnings. ++ on volatile variable, arithmetic on dif enums, start change Location to Position

+++ src/game/BattleGround/BattleGroundAV.h
@@ -628,7 +628,7 @@ class BattleGroundAV : public BattleGround
         void ChangeMineOwner(AVMineIds mineId, PvpTeamIndex newOwnerTeamIdx);
 
         // World state helpers
-        uint8 GetWorldStateType(uint8 state, PvpTeamIndex teamIdx) const { return teamIdx * BG_AV_MAX_STATES + state; }
+        uint8 GetWorldStateType(uint8 state, PvpTeamIndex teamIdx) const { return (uint8 )teamIdx * (uint8 )BG_AV_MAX_STATES + state; }
         void UpdateNodeWorldState(AVNodeIds node, uint32 newState);
 
         // Herald announcements
